requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
greenlet==3.3.0
idna==3.11
lxml==6.0.2
numpy==2.4.0
playwright==1.57.0
pyee==13.0.0
pygame==2.6.1
python-pygaze==0.7.6
requests==2.32.5
setuptools==80.9.0
typing-extensions==4.15.0
urllib3==2.6.2
